#310-autoconfiguration: naming cleanup, add order to factStore
