+++ factcast-examples/factcast-example-client-spring-boot2/pom.xml
@@ -49,5 +49,9 @@
           <version>2.0.20.Final</version>
         </dependency>
     -->
+    <dependency>
+      <groupId>org.factcast</groupId>
+      <artifactId>factcast-store-inmem</artifactId>
+    </dependency>
   </dependencies>
 </project>
