+++ factcast-examples/factcast-example-client-spring-boot2/src/main/resources/application.properties
@@ -1,3 +1,5 @@
 grpc.client.factstore.negotiation_type=PLAINTEXT
 grpc.client.factstore.enable_keep_alive=true
 grpc.client.factstore.keep_alive_without_calls=true
+
+#spring.main.allow-bean-definition-overriding: true
