+++ factcast-spring-boot-autoconfigure/src/main/java/org/factcast/spring/boot/autoconfigure/server/grpc/FactCastGrpcServerAutoConfiguration.java
@@ -13,30 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.factcast.spring.boot.autoconfigure.client.grpc;
+package org.factcast.spring.boot.autoconfigure.server.grpc;
 
-import org.factcast.client.grpc.GrpcFactStore;
-import org.factcast.core.store.FactStore;
+import org.factcast.server.grpc.FactStoreGrpcService;
+import org.factcast.spring.boot.autoconfigure.store.inmem.InMemFactStoreAutoConfiguration;
+import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
-import net.devh.springboot.autoconfigure.grpc.client.AddressChannelFactory;
-
-/**
- * Provides a GrpcFactStore as a FactStore implementation.
- *
- * @author uwe.schaefer@mercateo.com
- */
-
 @Configuration
-@ConditionalOnClass(GrpcFactStore.class)
-public class GrpcFactStoreConfiguration {
-
-    @Bean
-    @ConditionalOnMissingBean
-    public FactStore factStore(AddressChannelFactory af) {
-        return new GrpcFactStore(af);
-    }
+@ComponentScan(basePackageClasses = FactStoreGrpcService.class)
+@ConditionalOnClass(FactStoreGrpcService.class)
+@AutoConfigureAfter(InMemFactStoreAutoConfiguration.class)
+public class FactCastGrpcServerAutoConfiguration {
 }
