+++ factcast-spring-boot-autoconfigure/src/main/java/org/factcast/spring/boot/autoconfigure/store/pgsql/PGFactStoreAutoConfiguration.java
@@ -15,8 +15,10 @@
  */
 package org.factcast.spring.boot.autoconfigure.store.pgsql;
 
+import org.factcast.spring.boot.autoconfigure.store.inmem.InMemFactStoreAutoConfiguration;
 import org.factcast.store.pgsql.PGConfigurationProperties;
 import org.factcast.store.pgsql.internal.PGFactStoreInternalConfiguration;
+import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.ComponentScan;
@@ -36,5 +38,6 @@ import org.springframework.context.annotation.Import;
 @ConditionalOnClass(PGFactStoreInternalConfiguration.class)
 @Import(PGFactStoreInternalConfiguration.class)
 @ComponentScan(basePackageClasses = PGConfigurationProperties.class)
-public class PGFactStoreConfiguration {
+@AutoConfigureAfter(InMemFactStoreAutoConfiguration.class)
+public class PGFactStoreAutoConfiguration {
 }
