+++ factcast-spring-boot-autoconfigure/src/main/java/org/factcast/spring/boot/autoconfigure/store/pgsql/rds/RdsAutoConfiguration.java
@@ -13,15 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.factcast.spring.boot.autoconfigure.server.grpc;
+package org.factcast.spring.boot.autoconfigure.store.pgsql.rds;
 
-import org.factcast.server.grpc.FactStoreGrpcService;
+import org.factcast.store.pgsql.rds.RdsDataSourceFactoryBeanPostProcessor;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.env.Environment;
 
 @Configuration
-@ComponentScan(basePackageClasses = FactStoreGrpcService.class)
-@ConditionalOnClass(FactStoreGrpcService.class)
-public class FactCastGrpcServerConfiguration {
+@ConditionalOnClass(RdsDataSourceFactoryBeanPostProcessor.class)
+public class RdsAutoConfiguration {
+
+    @Bean
+    public RdsDataSourceFactoryBeanPostProcessor rdsDataSourceFactorBeanPostProcessor(
+            Environment env) {
+        return new RdsDataSourceFactoryBeanPostProcessor(env);
+    }
 }
