--- factcast-spring-boot-autoconfigure/src/main/java/org/factcast/spring/boot/autoconfigure/store/pgsql/rds/RdsConfiguration.java
@@ -1,33 +1,0 @@
-/*
- * Copyright © 2018 Mercateo AG (http://www.mercateo.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.factcast.spring.boot.autoconfigure.store.pgsql.rds;
-
-import org.factcast.store.pgsql.rds.RdsDataSourceFactoryBeanPostProcessor;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.env.Environment;
-
-@Configuration
-@ConditionalOnClass(RdsDataSourceFactoryBeanPostProcessor.class)
-public class RdsConfiguration {
-
-    @Bean
-    public RdsDataSourceFactoryBeanPostProcessor rdsDataSourceFactorBeanPostProcessor(
-            Environment env) {
-        return new RdsDataSourceFactoryBeanPostProcessor(env);
-    }
-}
