+++ factcast-spring-boot-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -4,7 +4,7 @@ org.factcast.spring.boot.autoconfigure.client.cache.infinispan.FactCastInfinispa
 org.factcast.spring.boot.autoconfigure.core.FactCastAutoConfiguration,\
 org.factcast.spring.boot.autoconfigure.store.inmem.InMemFactStoreAutoConfiguration,\
 org.factcast.spring.boot.autoconfigure.metrics.MetricsAutoConfiguration,\
-org.factcast.spring.boot.autoconfigure.client.grpc.GrpcFactStoreConfiguration,\
-org.factcast.spring.boot.autoconfigure.server.grpc.FactCastGrpcServerConfiguration,\
-org.factcast.spring.boot.autoconfigure.store.pgsql.rds.RdsConfiguration,\
-org.factcast.spring.boot.autoconfigure.store.pgsql.PGFactStoreConfiguration\
+org.factcast.spring.boot.autoconfigure.client.grpc.GrpcFactStoreAutoConfiguration,\
+org.factcast.spring.boot.autoconfigure.server.grpc.FactCastGrpcServerAutoConfiguration,\
+org.factcast.spring.boot.autoconfigure.store.pgsql.rds.RdsAutoConfiguration,\
+org.factcast.spring.boot.autoconfigure.store.pgsql.PGFactStoreAutoConfiguration\
